Fix focus chance bug
Previously if a player did a focusable activity for longer than the
maximum amount of time, they would get a message saying that they can't
focus for that long and then possibly a message saying that they loose
track of time and focus for the max length.  Now they get one or the
other.